requirements.txt
apex==0.9.10.dev0
boto3==1.17.40
botocore==1.20.40
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
Cython==0.29.21
dataclasses==0.8
-e git+https://github.com/pytorch/fairseq.git@698e3b91ffa832c286c48035bdff78238b0de8ae#egg=fairseq
filelock==3.0.12
future==0.18.2
gdown==3.12.2
idna==2.10
importlib-metadata==3.7.0
javalang==0.13.0
jmespath==0.10.0
joblib==1.0.1
mkl-fft==1.2.0
mkl-random==1.2.1
mkl-service==2.3.0
nltk==3.6.1
numpy
olefile==0.46
packaging==20.9
Pillow
protobuf==3.14.0
pyarrow==2.0.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.6
sacrebleu==1.4.14
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.5.4
sentencepiece==0.1.94
six
tensorboardX==2.1
threadpoolctl==2.1.0
tokenizers==0.5.2
torch==1.5.1
torchaudio==0.5.0
torchvision==0.6.0
tqdm==4.55.1
transformers==2.6.0
tree-sitter==0.2.1
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0